[hotfix] Introduce common interface to all IncrementalKeyedStateHandles
